move test to a different file
